Revert "refactor: changed the database of mysql to postgresql"
This reverts commit 2d368dcc583ed549112d931435fe78a0a78b5550.

+++ src/entity/AmountToPay.ts
@@ -24,7 +24,7 @@ export default class AmountToPay {
     @Column('text')
     description: string;
 
-    @Column('float')
+    @Column('double')
     value: number;
 
     @ManyToOne(() => User, (user) => user.amountToPay, { onDelete: 'CASCADE' })
